+++ .idea/compiler.xml
@@ -2,21 +2,15 @@
 <project version="4">
   <component name="CompilerConfiguration">
     <annotationProcessing>
+      <profile default="true" name="Default" enabled="true" />
       <profile name="Gradle Imported" enabled="true">
         <outputRelativeToContentRoot value="true" />
         <processorPath useClasspath="false">
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.28/a2ff5da8bcd8b1b26f36b806ced63213362c6dcc/lombok-1.18.28.jar" />
         </processorPath>
-        <module name="CareHeim-Server.main" />
+        <module name="CareHeim.main" />
       </profile>
     </annotationProcessing>
-    <bytecodeTargetLevel>
-      <module name="CareHeim-server" target="17" />
-      <module name="CareHeim-server.main" target="17" />
-      <module name="CareHeim-server.test" target="17" />
-      <module name="com.spring.CareHeim-BE" target="17" />
-      <module name="com.spring.CareHeim-BE.main" target="17" />
-      <module name="com.spring.CareHeim-BE.test" target="17" />
-    </bytecodeTargetLevel>
+    <bytecodeTargetLevel target="17" />
   </component>
 </project>
