+++ .idea/modules.xml
@@ -3,7 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/CareHeim-BE.iml" filepath="$PROJECT_DIR$/CareHeim-BE.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/CareHeim-Server.main.iml" filepath="$PROJECT_DIR$/.idea/modules/CareHeim-Server.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/CareHeim.main.iml" filepath="$PROJECT_DIR$/.idea/modules/CareHeim.main.iml" />
     </modules>
   </component>
 </project>
