--- CareHeim-Server/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'CareHeim-Server'
