--- CareHeim-Server/src/test/java/com/spring/careheimserver/CareHeimServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.spring.careheimserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CareHeimServerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
