+++ CareHeim/src/main/java/com/spring/careHeim/CareHeimServerApplication.java
@@ -1,11 +1,12 @@
-package com.spring.careheimserver;
+package com.spring.careHeim;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.mongodb.config.EnableMongoAuditing;
 
 @EnableMongoAuditing
+@EnableJpaAuditing
 @SpringBootApplication
 public class CareHeimServerApplication {
