+++ CareHeim/src/test/java/com/spring/careHeim/CareHeimServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.spring.careHeim;
+
+import com.spring.careHeim.domain.users.UserRepository;
+import com.spring.careHeim.domain.users.entity.User;
+import org.bson.types.ObjectId;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CareHeimServerApplicationTests {
+
+}
